use a trait instead of duplicated code.
